Move config file parsing into its own method
This change splits handling of the VooDriver configuration file into its
own method, readConfigFile.

There is currently what I believe is a bug with the interaction between
items read from the config file and those passed in on the command
line.  One would expect that the order of precedence would be builtin =>
config file => command line.  That is, anything specified on the command
line should override what is in the config file.  Currently the config
file has precedence over the command line.  This will be fixed in a small
follow-on change.

Testing done:

Compiled both normal and debug.  Ran through javadoc to check the doc
format.  Created default and custom config files and checked that the
default is used unless --config is specified and that, with no config
file, the internal defaults are used.  Checked interaction with
command line vars (see above comment).