Add --haltOnFailure command line option
This change adds the command line option --haltOnFailure to VooDooDriver.  This
option, when specified, causes VooDooDriver to terminate the test run as soon
as a failing test is encountered.  The intent is mostly to make testing and
debugging VooDooDriver easier - I don't expect it will find much use in actual
testing.

In addition to the change, an expected-to-fail test suite has been added to
VooDooDriver's tests.

Testing done:

Ran Fail.xml suite with and without --haltOnFailure.  With option specified,
VooDooDriver terminates as expected.  Ran VooDooDriver.xml suite.