Bug 49253 - handle truncated log files
This change fixes bug 49253.  If a log file is truncated, the missing end tags
cause the XML parsing engine to throw an exception.  Should such an exception
occur, VDDReporter will now attempt to repair the XML by adding the appropriate
end tags and then will re-parse the file.  If a further error occurs, the
file is skipped and an error is printed.  If parsing is successful, the
file is highlighted in the results summary by setting the background of the
row to the color requested in the bug ("Orange 4" or #ffcc99).  If the log
file was simply empty, the name of the row is set to the name of the log file.

Testing done:

Ran against the provided results directory.  Both truncated files were parsed
successfully.  Created an empty log file and parsed it successfully.  Checked
javadoc comments in output documentation.