Add plugin cmdopt to VooDooDriver
plugin seems a reasonable thing to want to put into a soda-config.xml,
particularly considering that SugarWait is very nearly required.  This change
adds a cmdopt for plugin to VDD, then.  Also in the change is a slight
reworking of cmdopt parsing that reduces the amount of code needed to read
cmdopts.  To add a new one now, simply add the appropriate string to
the validCmdopts array.

Testing done:

Ran VDD with some cmdopts in the config file and checked that they took effect.