Defer cast to int in arg processing
This change defers the cast from a string to an int for two command line
options: --attachtimeout and --restartcount.  The reason for this is so that
argument parsing from both the command line and from the config file may
be further simplified.  There is no reason to have the cast take place in
two separate places.

Testing done:

Ran VDD with --attachtimeout and --restartcount on the command line and with
<cmdopt name="attachtimeout">... in the config file ("restartcount" is not
yet supported in the config file).