Fix a NullPointerException in VDDReporter
If a <testfile> entry in a log file contains a file with no path information
(e.g. foo.xml), VDDReporter would get a NullPointerException when it attempted
to use the filename's path (which was null).  This change fixes that by
checking for null and acting accordingly.

Testing done:

Ran VDDReporter against a log file that had been causing this exception.  No
more exception.