Add <input> event
This change adds an <input> event to VooDooDriver, which will match any <input>
tag in the web page.  The functionality is currently limited to functionality
common to all types of inputs.  Should there be a demand for <input> to have
all-encompassing functionality (including, for example, click and selected
actions), then those will be added at that time.

Testing done:

Ran a script with <input> against a page with <input> tags and made sure they
were found appropriately.