Follow on change to SodaElements
A minor change to SodaElements.xml didn't make it into the last commit.  This
change fixes that.

By way of explanation, since this particular minor change is a little odd,
the <button> Soda event is used all over the place to mean both
<button type="button|submit|reset"> and <input type="button|submit|reset">.
Strictly speaking, the latter is almost always what is meant, but to
workaround the enormous quantity of erroneous <button> event tags in the
code base, vdd will, with the follow-on change, allow both <button> and
<input> HTML tags, as long as the type matches one of the three button types.

Testing done:

Ran against a test script with some printfs sprinkled about to check the tag.