Fix bug 49533
The act of selecting a <select> element can trigger a page refresh.  Should
such a refresh occur, the remainder of the select event after the selecting
is done will be working with a stale element, and will cause a
StaleElementReferenceException to be thrown.

This change, then, checks whether the element is stale after selecting a
<select> element.  If it is, the element is refreshed, and execution can
continue.

Testing done:

Ran the failing script from the bug, Leads_1969.xml.  Before this change, a
StaleElementReferenceException occurred.  After, the "refreshing stale
element" log message appeared, and the exception no longer occurred.