Fix bug 49536
This change modifies the exception handlers for ElementNotVisibleException in
VDD to only log an error if the element is required.  If the element is not,
it is still noted that the exception was thrown, but as a normal log message
instead.

Testing done:

Ran the script from the bug.  Before an error message would appear.  After,
the log message was still there, but not as an error.