Fix a NullPointerException
This change fixes a NullPointerException that was introduced in the previous
change.  The reporter object in SodaXML is not always present, but the previous
change assumed it was.  This change checks for it before using it.

Note: It would be worth doing to straighten out the handling of this reporter
object so that *something* is always there, even if it's not *the* reporter
object.

Testing done:

Ran several scripts from the SodaIceBox suite (google.xml was the one running
when the exception was encountered first).