Finish moving VDD to package
This change updates all the VDD sources to be in either package
org.sugarcrm.voodoodriver or org.sugarcrm.vddlogger, depending on which it
belongs to.

After this change, external plugins (e.g. sugarwait) need to be updated to
point to the correct package.

Testing done:

Built (normal and debug) and ran against a few soda suites after fixing and
rebuilding sugarwait.