Bug 49536 (continued): Make error/log message more clear
This change continues the previous change to address bug 49536 by adding more
information to the message printed out when an element is not visible.  If
required is false, that is printed.  In all cases, the methods used to select
this element are printed.

Testing done:

Ran a test script with some non-visible elements, made sure the message printed
is appropriate.