Bug 49306: Fix StaleElementReferenceException in <image> event
This change slightly rearranges some code so that clicking an img no longer
references it before the end of the event.  The way the code was before, the
handleVars call occurred after clicking, which, if the click caused the page
to refresh, would throw this exception.  Now the call to click() is the last
item in the event.

Testing done:

Ran test script from the bug with no errors.