Fix a NullPointerException
This change fixes SodaBrowser.executeJS() to avoid a NullPointerException. Null
is a valid return value when executing scripts, although it is rarely (if
ever?) seen on firefox. Chrome does return null, thus triggering this
exception. This method was restructured to check for null and only call
toString if the object is not null. Some minor code and document cleanup was
also done.

Testing done:

Ran against a script that had been throwing NullPointerExceptions.