Add more structure to voodoo.texi
This change simply adds more document structure to the new VDD documentation.
The largest unstructured section deals with the HTML events.  I expect this
to be moderately unpleasant to put into order (*which* order being a prominent
question).

Testing done:

Created .html, .info, and .pdf files and verified the output is more or less
as desired.