Refactor VooDooDrivers configuration handling
This change refactors the way VDD handles configuration data coming from
config files and the command line.  A long-standing bug in which config
data from the config file overrides config data from the command line is
fixed.  It begins the process of passing around a single config structure,
rather than multiple variables.  This will greatly simplify future additions
to the codebase.

There are still several follow-on changes to this that should results in
a more robust and easier to add features to VDD.

Testing done:

Ran with various combinations of command line and config file options.  Ran
VDD in the same manner as jenkins and verified no change to functionality.