Rename Elements.xml to Events.xml
This change, the first in a series, begins rearchitecting the way VDD Events
are handled internally.  In this change, the file Elements.xml is renamed to
Events.xml.  In VDD terminology, "elements" are HTML elements found on an
HTML page; "events" are the smallest unit of VDD test execution and may
correspond to an element, but may instead correspond to an action taken by
VDD (e.g. "tr" is an event that refers to a "tr" element, but "csv" is only
an event).

A very small amount of cleanup was also done within Types.java.

Testing done:

Ran VDD to make sure Events.xml is correctly loaded.