Add methods for firing plugins without element/plugin event specs
This change adds two firePlugin methods:

   boolean firePlugin(WebElement element);
   boolean firePlugin(WebElement element, PluginEvent event);

They are intended for use when plugin firing is not dependent on the current
HTML element (or there is no current HTML element); or for when there is
neither a current HTML element, nor is there a need to filter by plugin
event type.  Both routines, as with the original three-argument firePlugin
fire all plugins matching the provided specifications and return true if
all plugins executed successfully.

Testing done:

Ran VDD against a Soda suite.  These routines aren't yet used, so their
testing was limited to compilation.