Add auto-screenshot functionality to VDD
This change adds functionality to VDD to automatically take screenshots when
certain events occur.  The events, which can be specified either on the
command line with the --screenshot argument or within the VDD config file
with a <cmdopt name="screenshot">, are warning, error, assertfail, exception,
watchdog, and all.  Screenshots are dumped into a test-specific "screenshots"
directory.

Since much of this functionality parallels --savehtml, the HTML page saving
functionality was slightly rewritten so that common code could be factored
out.

Testing done:

Ran VDD against its regression test and against some Soda suites using various
combinations of --savehtml and --screenshot events.  Verified that saved HTML
pages and screenshots are where and as expected.