Clean up and document Browser and its subclasses.
This change cleans up the Browser class and its three subclasses and adds
documentation to all methods and fields.  In addition, the BrowserInterface
interface is removed since the abstract class Browser performs all the needed
functions that would be provided by having an interface.

Testing done:

Compiled VDD, checked the documentation, ran the VDD regression suite, and ran
a couple of Soda suites.