Rework slightly VDD-specific exceptions.
This change adds a new VDD-specific exception, VDDException.  This is intended
to wrap lower-level exceptions for propogation up the call chain for
situations that can not be handled at a lower level.  The PluginException
class is also reworked slightly to be a child of this new class.  Additionally,
a minor change was made to code that uses PluginException in response to
testing done.

Testing done:

Ran VDD, using an invalid plugin file to force an exception.