Add plugin access to Soda Vars
This change adds getSodaVars and setSodaVars methods to PluginData.  With this
change, plugins now have access to the values of Soda Vars.  The VDDHash
containing Soda Vars passed to the plugin is a shallow copy -- as long as
no complex data structures are added to the collection of Soda Vars (an
unlikely occurrence: they are all strings), this means the plugin can do as
it pleases with its Soda Vars without affecting VDD.

Testing done:

VDD regression suite with added code to print Soda Vars from within the plugin.