Bugs 51806 and 51807: Fix <textarea> attributes
This change modifies the set attribute of the textarea event to first clear the
text in the textarea before adding the new text.  An append attribute is also
added, providing the old functionality, but with a more sensible name.

Testing done:

Ran VDD against a test script with a textarea.