Bug 49296: Add assertselected attribute to <option>
This change adds a new assertion to the option event, enabling a check of
whether any or no options are selected.  For single item selects, there is
always one selected option, so this assertion is meaningless.  On multi-item
selects, however, this can be used to assert that nothing is selected.

Testing done:

Ran against the script from the bug and against an expanded script with more
variations.