Create VDD log file.
This change creates a mechanism so that all of VDD's output can be logged to
a file.  System.out and System.err are replaced by a new object, VDDLog that
outputs both to the console and to the new voodoo.log file.  This should aid
with some debugging tasks and will keep a nice record of all of VDD's config
which currently is printed to the console, but not saved anywhere.

Testing done:

Ran the VDD test suite and tail -f'd the log file in another terminal.