Allow multiple plugins
This change modifies VDD's command line plugin logic to allow multiple plugins
to be specified on the command line and in its config file.  There was no
runtime limitation, but the plumbing didn't yet exist to move arrays of
plugins from VooDooDriver.main and friends down to EventLoop.  This change
adds that plumbing.  Some cleanup remains, but it isn't just for --plugin, so
it will wait for a little while (this cleanup is pushing all config file
processing down to Config.java and using options.xml for both command line and
config file specifications).

Testing done:

Ran VDD with multiple plugins and one and none.  All scenarios worked.