Load Events.xml in Event's static block
This change adds code to load Events.xml in the static block in Event.  The
type created from Events.xml is now ArrayList<VDDHash> instead of the wrapper
around that type, ElementsList.  That class will be removed in a future change.
With the loading done in a static block, this file will be loaded only once
during an entire VDD run, instead of once per test.