Add a big, ugly Event Enum to Event.java
This is an interim step on the way to VDD code niceness.  While this sort
of large list is exactly what I'd like to get rid of (it merely duplicates
the same list in Events.xml), navigating the treacherous waters of the class
loader to load Event subclasses based on the classname being specified per
event in Events.xml will take more time now than is worth it, and so that
task is deferred until more code elsewhere is in place.