Create Event constructor
This change gives the Event superclass a constructor.  All subclass will call
it.  Its purpose is to take the event specification in the test file, and
compile it down into an internal list of selector and action attributes that
will then be used to find the appropriate HTML elements (selectors) or to
run voodoo on them (actions).