Create <browser> event
This change moves the browser event code out of EventLoop and into its own
event.  The code was somewhat simplified as well.  Some slight reworking of
the Event base class and the code in EventLoop that runs event was needed.

A unit test script for <browser>, browser.xml, was added to the VDD test suite.

Testing done:

Ran the VDD test suite.