Create the <attach> event.
This change creates the Attach event object.  The attach event attaches to
another browser window, then executes child events in that window, and
finally switches back to the parent window.

Testing done:

Compiled.  Since the new link event hasn't yet been implemented, it's not
possible to create a popup window, and so it's not possible to adequately
test this.  Soon, though...