Create HtmlEvent
This class is the parent class of all VDD HTML events (e.g. link, div, etc.).
Most of the functionality scattered about (and duplicated dozens of times in)
EventLoop will end up here.

Within this class are several nested classes that are used for action
execution.  Each action that can be done by an event is represented by a
nested class within HtmlEvent that implements the nested interface Action.
Some of the other nested classes aren't actions, but are used for action
bookkeeping.

Testing done:

Just compiled (which went well).  Nothing is hooked up to be runnable yet.