Make search by value work with IE
VDD's current search by value routine uses the Selectors API to search for
elements with matching tags, types and values.  This works well when the
API is available.  However, on IE the API is only available with IE8 and
IE9 and then only when it is rendering the page in standards mode.  Since
there's no guarantee that any page will ever be rendered in standards mode
(it requires a special HTML tag to be present on the page), it's safe to
assume the Selectors API will never be available on IE.  With this in
mind, find by value has been modified for IE to use a brute force
javascript search.  This relies on the html_tag and html_type attributes
of events.  Additionally, the find by value using the Selectors API has
been modified to more fully use the html_tag and html_type information.
Now all elements that can have the value attribute will use the correct
search parameters.

One strange IE bug that probably isn't work fixing is that the following
HTML element:

   <button type="reset"/>

won't be found using this method.  It seems to be a bug in IE because
manually searching for it with the javascript console doesn't work either.
Should this become a problem, this will be revisited.

Also, IE is having difficulties with <input type="file"/>.  This will be
addressed seperately.

Testing done:

Ran values.xml.  All events pass both on IE and on Firefox, with the
exception mentioned above.