Bug 53577: allow window closing alerts
This change fixes a bug in the alert event that would allow a Javascript
exception to propogate when an alert was being handled that appeared in
response to a window closing event.  The switch back to the main window code
was causing a Javascript exception in the WebDriver server-side code. That
exception is now handled.

Testing done:

Ran failing test script.  Javascript exception is now handled and a suitably
benign message is logged.