Bug 53899: Fix --assertpagefile functionality
A change from a few months ago (the refactoring of Test's constructor) caused
the assertpagefile functionality to stop working.  The code that extracts the
filename from the config hash used the key "assertpage" instead of
"assertpagefile".  It didn't help that the same file is referred to as both
"assertpage" and "assertpagefile" in method and variable names throughout the
source, with as many different capitalizations as can be imagined.

Testing done:

Ran the test case from the bug.  It passed (or rather failed) as expected.