Bug 44048: Create <option> event
This change implements part of the requested functionality in bug 44048: the
option event.  A standard set of attributes was given to this event.

A strange wrinkle with the selectors API was found: for some reason, it seems
impossible to find an option event using its value attribute like so:

  option[value="foo"]

even when the corresponding HTML is:

  <select><option>foo</option></select>

The HTML 4.0.1 spec is clear that "foo" should be the value of the value
attribute, and inspecting the DOM shows that it is.  Strangely, though, using
a CSS selector for that doesn't work.

To cope with that, the iterative code used to find elements by value developed
for IE is used when no html_type attribute has been defined for an event.  So
far, that only means the option event.  That code works well and finds option
tags.

Testing done:

Ran a test script to exercise the new event's attributes.