Fix VDDReporter path stupidity.
In several places, VDDReporter blithely assumes the '/' character is used
to separate path components.  This is fine on Linux (and MacOS), but
rapidly breaks down on Windows.  Using only '/' is OK there, and using the
native separator '\' is also OK, but mixing the two is generally not.
This change fixes all of the instances of such breakage I could find.  It
also scratches the surface of cleaning up this program.  A more thorough
job could take quite some time and so many glaring sections in need are
left for the future.

Testing done:

Ran against an output directory of VDD.  Checked that all options still
work (even if --help prints nothing of use).