Fix file/path handling in suite files
This change fixes the largest area of path separator difficulties within
VDD: suite files.  Rather than passing strings around and manually
fiddling with '/' and '\' conversion, (almost) all files and paths are
converted to File objects and manipulation is done with those instead,
letting the Java lib handle all the details.  A goodly amount of cleanup
was done to SuiteParser as well.

The most major change, though, is the addition of a new gvar:
global.basedir.  Prior to this change, relative paths in suite files are
taken to be relative to whatever directory VDD started running in, rather
than the directory the suite file is in, which makes more intuitive sense.
To avoid mass breakage, this behavior is still the default.  However, if
the gvar "global.basedir" is set, it is used as the base directory for
relative paths in suite files.

Testing done:

Ran sugarinit.xml with and without global.basedir and with paths specified
with different combinations of '/' and '\' directory separators.