Work around a java.io.File "feature".
When creating a File object with both parent and child paths, the javadocs
state that, if the child is absolute, it will be resolved in a
system-dependent manner.  What this turns out to mean is that it will be
simply appended onto the end of parent.  Obviously, if the child path is
absolute, that won't work as expected.  This change works around that
problem by checking whether the child path is absolute and only appending
it to the parent path if it isn't.

Testing done:

Ran the problem case.