Plumb --haltOnFailure into Reporter
The one place in the code that is always run when an error occurs is
within the Reporter object.  So, to make --haltOnFailure actually halt on
failure, this change plumbs the option down into Reporter's routines for
error reporting.

Testing done:

Ran a test script with errors.  VDD terminated as expected.