Fix Z-order bug with input.clear()
This change works around Selenium issue 3402.  On IE, when a text field, a
password field, or a textarea gets cleared using WebElement.clear(), the
IE window is sent to the back of the Z-order.  This only happens the first
time clear() is called for any element.  Rather than wait for a bug fix,
this change uses a Javascript snippet on IE to clear the control.

Testing done:

Ran clear() on a text, password, and textarea element.  The window stayed
in the foreground.