Update selenium libs to 2.25.0
This change updates the selenium libs to the latest version, 2.25.0.  The bug
that had been blocking the upgrade, 3393, has been fixed for Firefox.  Fixes
for IE and Chrome are presumed to be forthcoming.  With this change, Firefox
up to 14 is confirmed to work.  A follow-on change implementing the new alert
functionality will be made.

Testing done:

VDD test suite.  Further testing will be done after the alert handling code
has been added in the next change.