Add unexpectedAlertBehavior capability to Firefox
This change folds in the new alert functionality from Selenium, returning
Selenium's functionality with respect to alerts to its pre-2.18 mode of
operation.  With this change, alerts will not be dismissed and an
UnexpectedAlertException will not be thrown when one is seen.  Rather, as
was the prior behavior, VDD scripts will continue to explicitly handle them.

Testing done:

Admin Soda suite.  All alerts were handled as expected (no hangs).  The few
errors were unrelated to this change and are similar to those seen when
running VDD master.