Bug 55786: Stop reloading select element
This change reworks the select event so that the actions that can cause page
loads are executed last.  This obviates the need to reload the element and
fixes both this bug and the original bug, 49533, that resulted in the reload.

As part of fixing this bug, the select event was rewritten significantly.
This will hopefully make the code more logical and easier to maintain.  In
addition, the unit test for the select event was rewritten to cover all
(positive) cases.

Testing done:

Ran the VDD select unit test.  VDD has no way to perform negative tests, so
that testing was done with temporary modifications to the unit test.