Create the whitelist event.
This change creates the whitelist event.  This event adds items to the
whitelist that is used to allow strings to occur that would otherwise be
failures during page assert processing.  As a part of this change, some amount
of work was done on the page assert machinery, though by no means all that
needs to be done.  Of note perhaps is that the pageAssert attribute is now
uniformly pageassert (all lower case).  This matches all other attributes.

Testing done:

Ran the VDD test suite.