Implement the javascript event.
This change implements the javascript event, an event that executes javascript
code in the context of the current browser window.  A minor change leftover
from the previous change was also tacked to this one.  One slight functional
difference is that the contents attribute is no longer deprecated.  After
discussion with interested parties, it seems there is no good reason to
deprecate that functionality, and many good reasons for its use.

Testing done:

VDD test suite.