Create the pluginloader event.
This change creates the pluginloader event, an event that dynamically loads
java plugins.  A change also was made to exception handling in the main event
loop: if an exception has a nested exception, it is unpacked and reported
instead of the outer exception.  Since all exceptions coming from event
execution are VDDExceptions around the real exception, this just makes more
sense.

Testing done:

VDD test suite.