Bug 56570: Replace text in 'exists' attribute.
This change makes variable replacement in the exists attribute.  During
execution of the link event, the text in this attribute is replaced in the
event itself and various log messages are made with this.  However, the code
in findElement takes the unreplaced text out of the event and uses that
instead, then, confusingly, prints out hardcoded text for the values so it
looks as though replacement is being done incorrectly.

Fortunately the fix for this mess is a one-liner.

Testing done:

Ran the test script from the bug.