Error out correctly with script error
If there is a syntax error in a test script loaded by the script event, VDD
doesn't check and then gets a NullPointerException when it tries to use the
(nonexistent) events from that script.  This change fixes that behavior to
check for null return from TestLoader.getEvents() and to log an error in that
case.

Testing done:

Ran VDD and loaded a bad test script from another test script.  Saw appropriate
error message.