Add message parameter to ReportException.
All over the place in VDD, when an exception occurs, first a description string
is logged as an error, and then the exception itself is logged.  This ends up
counting exceptions as both errors and exception, and potentially saving the
same page twice and taking two screenshots.  Silly behavior.

This change adds a message parameter to ReportException so these descriptions
can be made part of the exception reported and the extra error accounted can
go away.

Testing done:

Ran a script with an exception checked that it was no longer counted as an
error, too.