Implement the div event.
This change implements the div event.  Since the div event is identical to many
other HTML events (except for the name), a common class for these identical
events was created, SimpleHtmlEvent.  Should an HTML event be created that
needs special functionality, it will be made into its own class.

Some extra plumbing was needed for this change as well.  Most notably, the
hidden element exception is now handled for all HTML events; this necessitated
that the value of the required attribute be plumbed up into EventLoop.

Testing done:

VDD test suite.