Some cleanup in EventLoop.
Event with all the code getting moved into event classes, EventLoop is still
big and unwieldy.  This change cleans it up a bit and adds documentation.  All
methods and variables that will still be there when the work is done are now
documented.  All code that will be deleted is now consolidated and marked.

Testing done:

VDD test suite.