Make sure test loading exceptions go to the log
This change causes exceptions generated during test loading to be properly
unpacked and reported to the log.  Previously the actual exception was being
dumped to stderr, and what was reported in the log was just the wrapper
exception object -- not terribly useful.

Testing done:

Ran a script that would generate an exception.