Add retry to click action.
Clicks can sometimes fail when elements are out of the viewport (see Selenium
issue 2766).  This change attempts to scroll the element into the center of
the viewport should a click fail.  The click is then retried.

Testing done:

VDD regression suite.  Doesn't fully address the issue on Chrome, but is better
than before.