Add email event to VDD.
EMAIL is a new HTML element introduced as part of HTML5.  Very similar to
INPUT type=TEXT, it adds implicit checking of email format so that web
devs don't have to.  This change adds support for that element to VDD.

Testing done:

Simple test script.  When this gets moved to dev_v2, it'll get a full unit
test.