Rework PageAsserter
This change heavily modifies PageAsserter.  The impetus was the observation
that runtimes  using PageAsserter are noticeably slower than those not using
it.

With the search regexes optimized a run of sugarinit with the old code took
177s.  Without using PageAsserter, a similar run of sugarinit took 151s.
With the code in this change in use, the runtime was 169s.  This is
approximately a 30% reduction in runtime of this section of code.

A new class has been added to encapsulate the process of searching with either
a regular expression or plain text, TextFinder.  In a follow-on change, this
will be incorporated into Reporter.java which uses this idiom extensively.

Testing done:

Repeated runs using jvisualvm and sampling to determine runtime and effects of
code changes on performance.  Verification of VDD results from page assertions
before and after.  Read generated javadocs.