Remove '/' regex delimeters.
Many of the regexes used in the test suite include additional '/' characters
as delimeters around the entire regex.  As part of a previous change dealing
with regular expression support, the code that stripped these was removed.

This change adds back code to delete '/' at the beginning or end of a regular
expression.

Testing done:

Verified that regexes with the extra characters are passing again.