Bugs 57864 and 57865: Accept unhandled alerts
This change adds code that will accept an unhandled alert some time after it
has appeared.  While the test script should anticipate and handle alerts
itself, should this not be done, it is not acceptable for VDD to go off the
rails for the rest of the test run.

The three places where unhandled alerts have been seen to cause problems
are in javascript execution, browser action handling, and fetching the page
source.  All three places are handled now by a new method in Reporter,
unhandledAlert().

Testing done:

Ran Cases_1280.xml and ran different test script that had unhandled alerts
on entry into <javascript> and <browser>.