Change StopEventException's ancestor
This change switches StopEventException so that it descends from
RuntimeException rather than from VDDException.  The reasoning is primarily
so that an event's execution can be terminated due to a runtime error
(rather than an exception) without going through contortions or without having
to declare every method to throw StopEventException.

Testing Done:

VDD Suite.