Delete some unusable and unimplemented events.
This change removes references to four events - optgroup, option, fieldset,
and legend - from Events.xml.  Of the four, only legend represents an element
that can be visible, and that is not used in the test suite.  Should
it ever be needed, it can easily be added, but this eventually is not
anticipated.

Testing done:

Compiled VDD.