Harden VDD against unexpected alerts
WebDriver throws WebDriverException in response to "miscellaneous" errors at
runtime.  In many cases, these errors are caused by an unexpected alert.
While it would be very nice to check for an alert when such an exception is
received and the do the Right Thing WRT accepting or dismissing it, it's
impossible to predict what that Right Thing is, and so the issue is punted to
the user.  This change makes that punting a bit less severe, although there
is little that can be done about VDD going off the rails in response to such
a problem.

Testing done:

Ran a test script with inconveniently placed alerts.