Add "format" attribute to timestamp event
This change adds the optional "format" attribute to the timestamp event.  Using
this attribute allows arbitrary date/timestamp formats to be used.  For a
complete description of allowable formats, see the java 6 documentation at:
http://docs.oracle.com/javase/6/docs/api/java/text/SimpleDateFormat.html

Testing done:

Ran the following test script:

<soda>
   <timestamp/>
   <timestamp format="yyMMddHHmm"/>
   <timestamp format="zebra"/>
</soda>

The first two worked, and the third produced an error due to the invalid format.