Reopen browser window when loading url
If the browser window has closed and the browser event is called with the
url attribute, this change causes VDD to reopen the window and then load the
URL.

This functionality is not currently used by the test suite, but it is very
handy for VDD regression tests.

Testing done:

Many iterations of the new sugarwait test, which closes and opens the window
several times.