Create SugarWait regression suite
Since SugarWait is such a fundamental component of testing Sugar with VDD
and it has proven in the past to have complicated interactions with VDD,
Sugar, and Selenium, it just makes sense for it to have its own regression
test.

The test added in this change hits all the corner cases of SugarWait that
have proven troublesome and also the normal cases as well.

This test is not being made part of the regular VDD regression suite due to
its slightly more complicated setup than the rest of the suite (path to
the plugin file must be specified for this test only), and also because
VDD changes rarely impact SugarWait functionality.

Testing done:

Many iterations of this test against SugarWait with output compared to
what it should be.