Bug 60042: Give TextFinder another search method
Some clients of TextFinder expect to match the complete text, and others
expect to match within the text.  This change adds a findExact method for
those clients that want a complete match.  The find method searches for
matches within the text.

Testing done:

Ran against a test script that uses all clients of TextFinder: PageAsserter,
<attach>, and assert/assertnot.