Use the timeout value when finding elements
Only one of the three methods in the Browser class for finding elements was
being used and only from one call site.  Most calls to findElement were
calling selenium directly, and, as a result, not honoring the timeout value
specified in the event.

This change fixes that by creating a wrapper around selenium's findElements
in EventLoop and deleting the methods in Browser.

Testing done:

timeout-specific test.  VDD regression suite.

Conflicts:

	src/org/sugarcrm/voodoodriver/EventLoop.java