Dismiss unexpected alerts when closing browser
Though browser closure through the browser event has been able to handle
unexpected alerts for some time, the restart test machinery was missed
(since it's handled at the VooDooDriver level).

This change fixes that omission.  Also, since the FirefoxDriver doesn't seem
to actually throw UnexpectedAlertExceptions in the presence of unexpected
alerts (it throws a JavaScript Error(!)), the scope of exception handling
is expanded to WebDriverException.

Testing done:

Ran a test with one script leaving an alert open prior to running a
restart test.  Both firefox and chrome handle that now.