Append to voodoo.log
VDD's voodoo.log output file was being created anew with each run.  When
there's only one output directory per run, that is fine, but on jenkins,
the output directory is reused which meant that only the output of the last
run was being preserved.

This change adds the append flag to voodoo.log creation, ensuring that all
output is captured.

Testing done:

Ran VDD multiple times with the same output directory.