Add window maximize/restore support to VDD
This change adds to VDD the ability to maximize and restore the browser
window.  In order to do so, the browser event is used like so:

   <browser action="maximize"/>
   <browser action="restore"/>

In addition to this functionality, a new command line option and config file
cmdopt have been added, --maximizewindows and maximizewindows, respectively,
that, when set, automatically maximize all browser windows as they're
created and get the focus.

Testing done:

Ran test scripts that exercised the browser actions, and that tested
the auto-maximize of both primary windows and attached windows.