Bugs 61883 and 61884: Restore regex delims
During the overhaul of regular expression and text searching in VDD, an attempt
was made to normalize the handling of regular expression input by only
actually doing a regular expression search if the provided string is a
regular expression.  As part of that, the "/" delimiters in use were ignored
and stripped.  As it turns out, that was a bad idea since searches are done
using regular expression metacharacters in the expectation that they'll be
a plaintext search (bug 61883).

The fix is simple: in order to be considered a regular expression, a provided
string must both contain metacharacters and be delimited by "/".  The real
strangeness, that the XML tag for both text and regex searches is <regex> is
left alone.

Testing done:

Various searches with various strings.  Correct behavior seen in all cases.